backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
agno==0.1.0
exa-py==1.0.11
google-generativeai==0.3.2
